update vm props to int64

+++ model/vm.go
@@ -10,6 +10,6 @@ type VM struct {
 	Name   string    `sql:"name,type:varchar(50)" json:"name"`
 	Status int       `sql:"status,type:int4" json:"status"`
 	HostID uuid.UUID `sql:"host_id,type:uuid" json:"host_id"`
-	VCPU   int       `sql:"vcpu,type:int4" json:"vcpu"`
-	Memory int       `sql:"memory,type:int4" json:"memory"`
+	VCPU   int64     `sql:"vcpu,type:int4" json:"vcpu"`
+	Memory int64     `sql:"memory,type:int4" json:"memory"`
 }
